Saving progress on bperft
bperft is aperft with directional slider attack tables.

Each square that is attacked by a slider knows which square the
attacking slider(s) are on, and those squares are stored according
to which direction the attacks are coming from.

The structure of the slider attack table is designed to make extending
and truncating the attacks quick and easy.  Currently the Exec and Undo
functions update these maps.  It *might* be faster to copy the maps from
parent to child and update them during Exec so that Undo doesn't need
to update the maps again.  That copy/update operation can be skipped
when there are no sliders on the board.

This work is not yet complete.  Additional code is needed to update
attack maps when en passant and castling moves occur.